update: 优化tofloat方法使其节省空间， close Air-duino/Arduino-AirMCU#32

+++ cores/AirMCU/WString.h
@@ -271,7 +271,7 @@ class String {
     // parsing/conversion
     long toInt(void) const;
     float toFloat(void) const;
-    double toDouble(void) const;
+    float toDouble(void) const;
 
   protected:
     char *buffer;         // the actual char array
